DocumentObject: simplify getSubObject() API
Changed dot separated SubName convention. The dot separated reference
inside SubName now demands and ending dot for any object name reference
inside, even if there is no non-object-element following.

The effect is that it now become apparent which part of the SubName
refers to object (anything before the last dot), and which for
sub-element (anything after the last dot, if there is any). If no dot
inside Subname, it means either no Subname, which means, it refers to
the whole owner object, or SubName is equal to the
non-sub-object-element name, such as Face1, Edge2, etc.

And because of this, getSubObject() API is simplified, such that it no
longer returns the subelement name like it used to be.