DocumentObject: add getSubObjects() API
DocumentObject::getSubObjects() returns name references of all child
sub-objects. It can be considered as a counter-part of
ViewProvider::claimChildren in App namespace, except that it returns
text name references instead of the children objects. The function is
exposed to Python and overridable with FeaturePython and
DocumentObjectExtension.

The default implementation searches PropertyPlacement for
transformation, and PropertyLink or PropertyLinkList for children.

This commit aslo adds the same default implementation of getSubObject(),
which means that any document object having a placement can now act like
a coordinate system (coordinate references), and we can use getSubObject
to obtain the the arbitarily nested transformation.

Also add support for using Label in getSubObject, buy prefix the name
reference with '$'