DocumentObject: add default impelementation of getSubObject()
The default implementation searches PropertyPlacement for
transformation, and PropertyLink or PropertyLinkList for children.

This means that any document object having a placement can now act like
a coordinate system (coordinate references), and we can use getSubObject
to obtain the the arbitarily nested transformation.

Also add support for using Label in getSubObject, buy prefix the name
reference with '$'