Added App::Link and Gui::ViewProviderLink
App::Link is an build-in link type object that links to other type of FC
objects with minimum overhead. It shares both the geometry modeling
data and 3D visual representation. It has built-in support of array
element for easy duplicating of the same object with independent
placement, material, and visibility.

The link capability is also exposed to Python with LinkBaseExtension
and LinkExtension, which makes it possible create your Python object
with customized link capability.