DocumentObject: bring visibility status into App namespace
A property named Visibility is added to DocumentObject which is kept in
sync with ViewProviderDocumentObject::Visibility. This effectively
brings the visibility status of the object into App namespace.

One practical usage, which is include in this commit, is that
Part.getShape can now construct compound from App::Part that honours the
child visibility status.